arch: consolidate core/dialog
An imperatively-opened dialog is one module — service, host component and
contract token — and it had been cut in half across the core/ <-> ui/ boundary.
The service must import the component it creates; the component must import the
contract and the input-manager types. Split across two folders, that is a folder
cycle by construction, and the same shape repeated for the confirm prompt.

Per #541, a service and the component it creates imperatively are placed as one
unit. DialogService is one-per-app with no template of its own, so the unit is
core/. ConfirmDialogService takes translation keys and an appearance — window
.confirm with styling — so it joins the shell it wraps.

InputLayer already lives in core/, so both back-edges dissolve rather than being
re-pointed. Pure move: no API change, no new indirection, no behaviour change.
Every remaining src/app folder cycle now closes through the single
auth.interceptor -> SignInComponent edge, which has its own issue.
