+++ frontend/src/app/core/dialog/confirm/confirm-dialog.component.html
