+++ frontend/src/app/core/dialog/dialog-content.ts
