+++ frontend/src/app/core/dialog/dialog-context.ts
