+++ frontend/src/app/pages/game-view/game-view.component.ts
@@ -1,8 +1,8 @@
 import { ChangeDetectionStrategy, Component, computed, inject, OnInit, viewChild, ViewContainerRef } from '@angular/core';
+import { DialogComponent } from '@core/dialog/dialog.component';
 import { InputManagerService } from '@core/input-manager.service';
 import { TranslocoDirective } from '@jsverse/transloco';
 import { testGameState } from '@testing/test-game-state';
-import { DialogComponent } from '@ui/components/dialog/dialog.component';
 import { SettingsComponent } from '@ui/components/settings/settings.component';
 import { CardsHandComponent } from './components/cards-hand/cards-hand.component';
 import { GameHeaderComponent } from './components/game-header/game-header.component';
