+++ frontend/src/app/pages/main-menu/main-menu.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { AuthService, User } from '@core/auth.service';
-import { DialogService } from '@core/dialog.service';
+import { DialogService } from '@core/dialog/dialog.service';
 import { getTranslocoModule } from '@testing/transloco.testing';
 import { SIGN_IN_DIALOG_OPTIONS, SignInComponent } from '@ui/components/sign-in/sign-in.component';
 import { BehaviorSubject, EMPTY, Observable } from 'rxjs';
