+++ frontend/src/app/pages/main-menu/main-menu.component.ts
@@ -1,10 +1,10 @@
 import { ChangeDetectionStrategy, Component, computed, inject, Signal, viewChild } from '@angular/core';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { AuthService } from '@core/auth.service';
-import { DialogService } from '@core/dialog.service';
+import { DialogComponent } from '@core/dialog/dialog.component';
+import { DialogService } from '@core/dialog/dialog.service';
 import { TranslocoDirective } from '@jsverse/transloco';
 import { MenuItem } from '@pages/main-menu/menu-item';
-import { DialogComponent } from '@ui/components/dialog/dialog.component';
 import { SettingsComponent } from '@ui/components/settings/settings.component';
 import { SIGN_IN_DIALOG_OPTIONS, SignInComponent } from '@ui/components/sign-in/sign-in.component';
 import { MenuItemsListComponent } from './menu-items-list/menu-items-list.component';
