+++ frontend/src/app/ui/components/settings/account/account.component.spec.ts
@@ -2,8 +2,8 @@ import { provideZonelessChangeDetection } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { AuthService, User } from '@core/auth.service';
-import { ConfirmDialogService } from '@core/confirm-dialog.service';
-import { DialogService } from '@core/dialog.service';
+import { ConfirmDialogService } from '@core/dialog/confirm/confirm-dialog.service';
+import { DialogService } from '@core/dialog/dialog.service';
 import { getTranslocoModule } from '@testing/transloco.testing';
 import { CREDENTIALS_DIALOG_OPTIONS, CredentialsFormComponent } from '@ui/components/credentials-form/credentials-form.component';
 import { BehaviorSubject, EMPTY, of } from 'rxjs';
