+++ frontend/src/app/ui/components/settings/account/account.component.ts
@@ -1,8 +1,8 @@
 import { ChangeDetectionStrategy, Component, computed, DestroyRef, inject, output, Signal } from '@angular/core';
 import { takeUntilDestroyed, toSignal } from '@angular/core/rxjs-interop';
 import { AuthService } from '@core/auth.service';
-import { ConfirmDialogService } from '@core/confirm-dialog.service';
-import { DialogService } from '@core/dialog.service';
+import { ConfirmDialogService } from '@core/dialog/confirm/confirm-dialog.service';
+import { DialogService } from '@core/dialog/dialog.service';
 import { InputLayer } from '@core/input-manager.service';
 import { listNavigation } from '@core/list-navigation';
 import { TranslocoDirective } from '@jsverse/transloco';
