+++ frontend/src/app/ui/components/settings/settings.component.spec.ts
@@ -3,10 +3,10 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { provideZonelessChangeDetection } from '@angular/core';
 import { By } from '@angular/platform-browser';
 import { AuthService } from '@core/auth.service';
+import { AH_DIALOG_CONTEXT } from '@core/dialog/dialog-context';
 import { SettingsService } from '@core/settings/settings.service';
 import { provideUserPreferencesService, UserPreferences } from '@core/settings/user-preferences.service';
 import { getTranslocoModule } from '@testing/transloco.testing';
-import { AH_DIALOG_CONTEXT } from '@ui/components/dialog/dialog.context';
 import { of } from 'rxjs';
 import { vi } from 'vitest';
 import { AccountComponent } from './account/account.component';
