+++ frontend/src/app/ui/components/sign-in/sign-in.component.spec.ts
@@ -3,8 +3,8 @@ import { provideZonelessChangeDetection } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { AuthService, User } from '@core/auth.service';
+import { DialogComponent } from '@core/dialog/dialog.component';
 import { getTranslocoModule } from '@testing/transloco.testing';
-import { DialogComponent } from '@ui/components/dialog/dialog.component';
 import { Subject } from 'rxjs';
 import { vi } from 'vitest';
 import { SignInComponent } from './sign-in.component';
